+++ physical_ai_tools/CHANGELOG.rst
@@ -2,6 +2,20 @@
 Changelog for package physical_ai_tools
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.5.9 (2025-07-07)
+------------------
+* Use global ROS connection instead of multiple instances
+* Add proper cleanup for image streams to prevent accumulation
+* Remove unnecessary scrollbars in Chrome browser
+* Contributors: Kiwoong Park
+
+0.5.8 (2025-07-07)
+------------------
+* Applied Redux Toolkit for better state management
+* Added heartbeat status to the UI
+* Added heartbeat topic publishing to monitor alive status of Physical AI Server
+* Contributors: Kiwoong Park, Dongyun Kim
+
 0.5.7 (2025-06-26)
 ------------------
 * Added Image Transport Plugin and fixed missing Gstreamer components